Revert "Move python modules under python/tensorflow_cloud/core"
