+++ tensorflow_cloud/machine_config.py
