+++ tensorflow_cloud/validate.py
