+++ tests/integration/call_run_on_script_with_keras_fit_cloud_build_test.py
