+++ tests/integration/call_run_on_script_with_keras_fit_test.py
@@ -35,18 +35,16 @@ import tensorflow_cloud as tfc
 
 # Automated MirroredStrategy: chief config with multiple GPUs
 tfc.run(
-    entry_point="tests/testdata/mnist_example_using_fit.py",
-    distribution_strategy="auto",
-    requirements_txt="tests/testdata/requirements.txt",
+    entry_point='tests/testdata/mnist_example_using_fit.py',
+    distribution_strategy='auto',
+    requirements_txt='tests/testdata/requirements.txt',
     chief_config=tfc.MachineConfig(
-        cpu_cores=8,
-        memory=30,
-        accelerator_type=tfc.AcceleratorType.NVIDIA_TESLA_T4,
-        accelerator_count=2,
-    ),
+            cpu_cores=8,
+            memory=30,
+            accelerator_type=tfc.AcceleratorType.NVIDIA_TESLA_T4,
+            accelerator_count=2),
     worker_count=0,
-    stream_logs=True,
-)
+    stream_logs=True)
 
 # Automated TPUStrategy
 # tfc.run(
