+++ tests/tensorflow_cloud/preprocess_test.py
@@ -16,21 +16,17 @@
 import os
 import unittest
 
-from tensorflow_cloud.core import machine_config
-from tensorflow_cloud.core import preprocess
+from tensorflow_cloud import machine_config
+from tensorflow_cloud import preprocess
 
 
 class TestPreprocess(unittest.TestCase):
     def setup_py(self):
-        self.entry_point = (
-            "python/tensorflow_cloud/tests/testdata/sample_compile_fit.py"
-        )
+        self.entry_point = "testdata/sample_compile_fit.py"
         _, self.entry_point_name = os.path.split(self.entry_point)
 
     def setup_ipython(self):
-        self.entry_point = (
-            "python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.ipynb"
-        )
+        self.entry_point = "tests/testdata/mnist_example_using_fit.ipynb"
         _, self.entry_point_name = os.path.split(self.entry_point)
 
     def get_preprocessed_entry_point(
